Moved classpath test to its own class
